chore: Moving to a publishable, per-use plugin structure
