+++ plugins/devx/index.ts
@@ -1,0 +1,1 @@
+export { markdownLogging } from "./markdown-logging";
