+++ plugins/sample/index.ts
@@ -1,0 +1,4 @@
+export { japaneseSEPlugin } from "./japanese-se";
+export { codeDetectionPlugin } from "./code-detection";
+export { mermaidPlugin } from "./mermaid";
+export { sillyPlugin } from "./silly-effects";
