+++ site/index.ts
@@ -11,7 +11,8 @@ import { PreviewRenderer } from "../src/commonmark/editor";
 import type { LinkPreviewProvider } from "../src/rich-text/plugins/link-preview";
 import type { ImageUploadOptions } from "../src/shared/prosemirror-plugins/image-upload";
 import { sleepAsync } from "../test/rich-text/test-helpers";
-import { devxPlugins, samplePlugins } from "../plugins";
+import { markdownLogging } from "../plugins/devx";
+import { codeDetectionPlugin, sillyPlugin, mermaidPlugin, japaneseSEPlugin } from "../plugins/sample";
 import "./site.css";
 
 function domReady(callback: (e: Event) => void) {
@@ -227,10 +228,19 @@ domReady(() => {
     const defaultEditor = getDefaultEditor();
     let plugins: EditorPlugin[] = [];
     if (enableSamplePlugin) {
-        plugins = [...plugins, ...samplePlugins];
+        plugins = [
+            ...plugins,
+            codeDetectionPlugin,
+            japaneseSEPlugin,
+            mermaidPlugin,
+            sillyPlugin
+        ];
     }
     if (enableDevxPlugin) {
-        plugins = [...plugins, ...devxPlugins];
+        plugins = [
+            ...plugins,
+            markdownLogging
+        ];
     }
 
     const options: StacksEditorOptions = {
