+++ tsconfig.json
@@ -5,5 +5,8 @@
         "outDir": "./dist/",
         "strictBindCallApply": true
     },
-    "include": ["src/**/*"]
+    "include": [
+        "src/**/*",
+        "plugins/**/*"
+    ]
 }
